OBP licensing at book-, chapter-, and image-level
An important point and two (silly?) questions:

How can we make clear in a machine readable way that it's necessary to go from the particular (individual image license) to the general (whole book)? - which is to say that the permission attached to the whole book excludes individual chapters, and the permission attached to individual chapters excludes images.

When the image is all rights reserved, which value should be given to the 'license-type' attribute?
And which URI should be included?